pyproject.toml
[project]
name = "officeqa-agentbeats"
version = "1.0.0"
description = "OfficeQA benchmark for AgentBeats competition - Green agent for evaluating document understanding"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    { name = "Arnav Singhvi", email = "arnav.singhvi@databricks.com" },
]
keywords = ["benchmark", "document-qa", "agentbeats", "evaluation", "llm"]

dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "httpx>=0.28.1",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
judge = [
    "a2a-sdk[http-server]>=0.3.20",
]
participant = [
    "a2a-sdk[http-server]>=0.3.20",
    "openai>=1.50.0",
    "anthropic>=0.40.0",
]
dev = [
    "mypy>=1.18.1",
    "pyyaml>=6.0.2",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["judge", "participant"]
